desktop/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "olympus-desktop"
version = "0.1.0"
description = "Olympus one-click desktop launcher (API + Worker + UI)"
requires-python = ">=3.10"
dependencies = [
  "pywebview>=5.1",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "uvicorn>=0.30.6",
  "psutil>=6.0.0",
]

[tool.setuptools.package-data]
"app.ui" = ["*.html", "*.js", "*.css"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
